tests: Fix buffer overflow
bytesFromFile() creates a std::vector that has exactly the size of the
file, but does not \0-terminate that. Passing this to functions that
expect a std::string implicitly calls the std::string(const CharT*)
constructor, which expects the buffer to be \0-terminated.

Fix the issue by calling the std::string(const Char*, size_type)
constructor, which correctly deals with strings that do not have
a terminating \0-byte.

Change-Id: Iaa676ebaff349216e79123814623eaa9c9e20e9d