mococrw: Fix leak in _X509_get_serialNumber
The ASN1 integers for 0 and maxLong created in _X509_get_serialNumber
have allocated memory associated with them and will lead to memory leaks
when not properly treated as such and freed.

ASN1_INTEGER objects on the stack could not be freed easily without
causing other errors, so switch to allocating them using
ASN1_INTEGER_new() entirely.

Change-Id: I21f394024f4bdbe82db800c899925e38e5df078f