mococrw: Add X509Cert::getSerialNumberDecimal
Provide a function that will return a decimal ASCII representation of
a certificate's serial number with arbitrary precision for use by RSU.

Note that this required adjusting template magic for the SSLDeleter
template, because BN_bn2dec returns a char*, but the corresponding
function to free the point is OPENSSL_free, which is a macro that
expands to a function that accepts void*. This would not compile with
the existing template.

Add a unit test that verifies correct functionality for both
getSerialNumber and getSerialNumberDecimal.

Change-Id: I41fef25f44c4886dc5a530ff39a81eca1f50014d