mococrw: Add getters for notBefore/After date as ASN.1
Since for example Necema uses ASN.1 Dates in its bindings that go up
to the year 9999, we need a representation for that. The library
until now converted the OpenSSL ASN.1 Dates to a time_point,
but this has a maximum value that is not big enough. Therefore we
had an overflow in the time_point in Necema.

We now add a call that returns the plain OpenSSL ASN.1 time.

Issue: MGUELVIS-24709
Change-Id: I836f985ae784accb79f584881b149d7799bf54a7