Do not generate symbols for inline functions
Specifying a function as inline but without static will in any case
generate a symbol and a non-inlined version of said function. This is in
most cases not what we want, if we want the function to be inlined. Add
the static keyword to all functions specified as inline to avoid this.

(cherry picked from commit d1b9c7cfcb8a496aa4a6ef206442195c4f04f0ee)