Add support for symmetric encryption
While MoCOCrW did provide reasonable coverage of common assymetric
cryptographic functionality and X.509 certificate management, it was
lacking a good C++ API to wrap symmetric encryption. Add support for
symmetric encryption using AES-CBC, AES-CTR and AES-GCM to address this.

Re-use the builder pattern that we have previously used for the X.509
functionality and provide an AESCipherBuilder class that can be used to
specify the various parameters required before creating the encryption
and/or decryption wrapper object.

Support both one-shot and streaming use-cases for symmetric
cryptography. Avoid unnecessary copies of the input and output buffers
when used with fixed block sizes in streaming en-/decryption.

For one-shot symmetric encryption using AES-GCM, the code is

  auto key = utility::cryptoRandomBytes();
  auto cipher = AESCipherBuilder(SymmetricCipherMode::GCM,
                                 SymmetricCipherPadding::PKCS,
                                 key)
      .buildAuthenticatedEncryptor();
  cipher->update(plaintext);
  auto ciphertext = cipher->finish();
  auto iv = cipher->getIV();
  auto authtag = cipher->getAuthTag();

In a similarly simple fashion, decryption can be done as a one-shot
approach:

  auto cipher = AESCipherBuilder(SymmetricCiphermode::GCM,
                                 SymmetricCipherPadding::PKCS,
                                 key)
      .setIV(iv)
      .buildAuthenticatedDecryptor();
  cipher->update(ciphertext);
  cipher->setAuthTag(authtag);

  std::vector<uint8_t> plaintext;
  try {
      plaintext = cipher->finish();
  } except (const MoCOCrWException& e) {
      std::cerr
        << "Decryption failed (possible auth tag validation failure?): "
        << e.what()
        << std::endl;
  }

Block-wise streaming encryption and/or decryption works by alternating
calls to update() and read()/readAll():

  auto cipher = ... /* as above */

  while (auto input = readInput()) {
      cipher->update(input);
      writeCiphertext(cipher->readAll());
  }
  writeCiphertext(cipher->finish());

and for decryption:

  auto cipher = ... /* as above */

  while (auto input = readInput()) {
      cipher->update(input);
      writePlaintext(cipher->readAll());
  }
  cipher->setAuthTag(authtag);
  try {
      writePlaintext(cipher->finish());
  } except (const MoCOCrWException& e) {
      std::cerr
        << "Decryption failed (possible auth tag validation failure?): "
        << e.what()
        << std::endl;
  }

The API attempts to automatically use sensible defaults where possible.
For example, it is not required to specify an IV for encryption. If no
IV is specified, one of the appropriate length for the chosen cipher
will be generated using a CSPRNG. Specifically, for AES-GCM, the
generated IV will be 12 bytes, for AES-CTR and AES-CBC, it will be 16
bytes long.