tests: Add new rootCA3 test set for CRL tests
The old test certificates used for the CRL testing have expired and the
private keys are no longer available. Set up a CA infrastructure similar
to tests/certs/rootCA{1,2} that should be used to test CRLs in the
future.

Generate a number of intermediate CAs and enduser certificates and
provide CRLs that mark those certificates and CAs as valid, revoked or
are expired.

Switch to SHA256 signatures for all CRLs over the SHA1 signatures that
were previously used. Ensure expiration dates for all certificates, CAs
and CRLs that we consider valid are 100 years into the future.

Extend the openssl.cnf used with a section for CRLs as outlined in [1].

[1] https://jamielinux.com/docs/openssl-certificate-authority/index.html