tests: Replace expired CRL test CAs and certs
Tests started failing because the root CA used to test CRL functionality
(as well as a few CRLs) expired beginning of 2020. Switch to a set of
certificates generated similar to our other test certificates where we
keep the keys around so we can prevent having to re-generate the entire
set next time (although the expiry dates are now all 100y into the
future, so this is unlikely to happen again during my and your
lifetime).