symmetric crypto: Allow encrypting empty plaintext
AES-GCM encryption fails in finalize() when update() was called with an
empty block while it attempts to obtain the auth tag (and with a useless
OpenSSL exception at that). This problem does not occur if we never call
EVP_CipherUpdate, so don't call the function when the given input block
is empty.

Add a test that verifies encrypting an empty block is possible in all
supported block cipher modes of operation. Adjust a test that would
otherwise fail with a false positive error.