tests: Fix failing tests when signatures are slow
We do specify the issuing time in the signing parameters explicitly, so
there is no need to check the issued certificate against the current
time with a two second precision limit. Using the expected value
directly and comparing without grace period prevents the test from
failing when it takes longer than 2 seconds.