Add workaround for OpenSSL PEM parsing bug
OpenSSL currently cannot parse PEM certificates that contain the
base64 encoded content in a single line and the length of the line
is a multiple of 254. Due to this, a workaround that splits the base64
encoded string into lines of 64 chars has been introduced.

OpenSSL Issue: https://github.com/openssl/openssl/issues/9187