ECDSA: Support verification from IEEE1363 format
ECDSA signatures are a tuple of two (large) unsigned integers
(r,s). Most libraries and protocols encode ECDSA signature in
DER serialized ASN.1 of SEQUENCE (int r, int s). This is also
what the current verifyDigest() and verifyMessage() functions
in the ECDSASignaturePublicKeyCtx expect.

In addition to this format, there is also a serialization format
specified in IEEE 1363. This serializes (r,s) in the form
padToKeyLength(r) || padToKeyLength(s). As this is binary
incompatible, it can't be processed with the current verification
functions.

This patch adds the verifyDigestIEEE1363() and verifyMessageIEEE1363()
functions that take an ECDSA signature in IEEE 1363 serialized form,
then convert it into the format that OpenSSL expects and then uses the
already provided functionality to validate the signature.