ECDSASignaturePublicKeyCtx: IEEE1362 support
Improve IEEE1363 support so that it is actually compatible
with the abstract verification context interfaces that are
implemented by ECDSASignaturePublicKeyCtx.

In order to do so, the constructor now offers to specify which
signature serialization format should be used by the verifyDigest()
and verifyMessage() functions (that are then also callable via
the abstract interfaces).

While touching this class, also clean up the Impl class so that
it now actually provides the implementation and not only data
storage with some sanitization.