Modify examples to be displayable on github and doxygen
The format of the examples is currently mixed. Some of the examples
are written to work with doxygen (and look not nice on github). Others
are written in the github md format (and thus look not nice in the
generated documentation). This patch modifies the examples to use
the same format (that works on github) and adds a script that
automatically converts the github md flavour into the doxygen md
flavour where necessary.