Update examples
Add examples for:
 - KDF
 - MAC
 - ECIES
 - EdDSA

In addition also improved the existing hash example (and some minor
fixes on the other existing examples) and reworked all the code examples
in a way that the functions/methods are clickable in the doxygen
documentation.

Also fixed the install target for the generated doxygen documentation.
Currently the install directory of the documentation is the build
directory of the documentation which leads to an endless recursion.