Update CI and build documentation in the CI
Build documentation in travis CI to fail the build if doxygen
detects error.

To do so, the docker container has been updated to ubuntu20.04
and doxygen has been installed. The Doxyfile.in has been updated
with doxygen -u on a ubuntu20.04 system.

Also removed the sudo:required entry as this seems to be not
required.