tests: Avoid random failure in stream cipher tests
Encrypting a one-byte plaintext with different IVs and keys using
AES-GCM or AES-CTR has a chance of (1/255)^2 to produce the same output
byte, which occasionally causes this test to fail. This is not a problem
for our block cipher tests, since they always produce outputs that have
a length that's a multiple of the block size.

Since our test does encryption with 0-byte, 1-byte and 10-byte strings,
filtering size() > 1 is enough to avoid this random failure.