Add support for AES-CMAC according to RFC 4493
This extends the existing generic `MessageAuthenticationCode` interface
with a second implementation besides HMAC. Given that the existing
naming suggests it this way, both implementations reside in `mac.cpp`
for now, but could be separated in future work.

The public interfaces are only extended, i.e. this change retains
API/ABI compatibility.

Co-authored-by: Clemens Lang <cl@clang.name>