Remove wrapper and mock for rsa_oaep_label getter
The way this function is called does not match the interface
offered by OpenSSL. It expects a pointer-pointer to unsigned char as
output argument, but we pass just a pointer. This probably went
unnoticed in the build due to SFINAE.

The wrappers for the function are not used anywhere in MoCOCrW, and are
also very unlikely to be needed by any of our users.