*PublicKeyCtx: Sig validation error message
In case that a signature is invalid, we just forward to error
message of the OpenSSLException generation by _EVP_PKEY_verify
to the MoCOCrWException. As an invalid signature is not really
a failure of the OpenSSL call, the OpenSSLException error message
is just error:00000000:lib(0):func(0):reason(0): 0.
This is very confusing and gives no indication that it actually
was the signature validation that failed.

This patch amends the error message and adds the part that signature
validation failed to the what() message of the MoCOCrWException
that is thrown.

Change-Id: Ibbe3e001fa554af9462d18a952d893d121c0e252