Replace Travis with new GitHub Actions CI workflow
The new pipeline uses an ad-hoc build of the Docker image providing the
building and testing environment. This removes the reliance on an
external Docker Hub account. A third-party layer caching component
helps to reduce the time spent on creating the image.

In addition to the previous pipeline, we now employ a matrix build to
check all combinations of clang/gcc and ASan enabled/disabled.