Do not rely on global openssl.cnf in CSR test
This invocation of the openssl tool tries to load the global
configuration file by default. Having the test be influenced by this
external artifact is not ideal. More importantly, it also fails if the
file does not exist. Removing this dependency will allow us to use the
much smaller "install_sw" target (no docs, no config) when adding
OpenSSL to the build container.