Fix CA tests
CA tests were failing randomly because CA's notBefore was set on
first use which then can cause problems in sanity check section of
CertificateAuthority::_signCSR function. If CA's notBefore was set
to +1 second compared to CA certificate's notBefore, it would
complain with "certificate is not valid yet" on the first
cert.verify(ctx). If CA's notBefore was set to -1 second compared
to issued cert, it would complain with "certificate expired" on
the second cert.verify(ctx).

Solution was to set CA's notBefore and CA certificate's notBefore
to the same value at the same time in test SetUp.

This commit also adds exceptions with better description if any
form of the above listed cases appear. This does not change
library behavior since this is how it worked before except exceptions
were made out of OpenSSL error messages which are less descriptive.