Adds main abstract classes related to HSMs (#116)
Adds two main abstract classes concerning HSMs. The HSM class is the highest abstract class in the hierarchy; all HSM implementations should inherit it either directly or indirectly. The second abstract class is the HsmEngine class which extends HSM and adds functionality based on OpenSSL's ENGINE_* API.

Since these two classes are both abstract, no concrete implementation of an HSM is provided in this PR. Such an implementation will be provided in a subsequent PR to ease review.