Fix front HSM interface (#117)
Converts the HSMEngine class, which makes use of OpenSSL's ENGINE API to interact with HSMs, from an abstract to a concrete class. Virtual functions which cannot be implemented due to current lack of support (e.g., key storage) are removed for now.

Adds readPrivateKeyFromHSM() and readPublicKeyFromHSM() to the MoCOCrW interface. Tests for loading keys are also included in this PR.

A simple code example is included to serve as a small integration test.