dilithium: add key-, signing- and verification-methods
As openssl is not supporting dilithium at the moment, the reference
implementation from the NIST PQC is used.

For this purpose the crystalls dilithium libraries without avx support
are linked.

The current version of MoCOCrW uses a EVP_PKEY centric implementation
for asymmetric cryptographic operations. To avoid changes on the
existing classes the relevant classes are cloned for the dilithium.

Functionality is added or removed wherever necessary. But the changes
were done with the plan in mind, to be able to change back to the
EVP_PKEY centric implementation once openssl supports dilithium.

The new dilithium feature is OPTIONAL.
To compile MoCOCrW with dilithium support specify -DDILITHIUM_ENABLED
when invoking cmake.
You have to make sure, that the static libraries created by libdilithium
(https://github.com/pq-crystals/dilithium) can be found the linker.
Additionally the compiler needs access to "api.h" from libdilithium.