hsm: add HSM key generation
HSM API is extended to support key generation on HSMs. This relies on patched
libp11 that supports key generation through OpenSSL ENGINE API. Loading of keys
is changed to support loading keys based on key id and key label. HsmEngine is
changed to associate it with a particular token so that keys can be generated
and loaded from multiple tokens.

As before, the point of entry for clients of this library who want to manage keys
is added to the key.h file

Co-authored-by: Tobias Kaufmann <Tobias.KA.Kaufmann@bmw.de>
Co-authored-by: John Galea <John.Galea@bmw.de>