CMAKE: Add CMake Header Configuration
Configures a header to define MACROs for optional features provided
by MoCOCrW. As a result, it avoids the need for users to add
target_compile_definitions commands to enable the features.

The approach taken uses configure_file() so that CMake generates
a config.h file from a template. The config file is included in
error.h, which in turn is used pretty much everywhere in the code
base.