HSM: Raise MoCOCrW Exception for unknown key ID
When attempting to load a key with an unknown ID,
MoCOCrW raises an OpenSSLException. This is not
ideal, as such exceptions are usually raised for
internal problems within OpenSSL. For cases where
an ID is unknown, one would expect a MoCOCrWException
to be thrown back to the user.

To make matters worse, the current catch-all
approach of using OpenSSLException makes it
difficult to distinguish different types of
errors.

In order to specifically identify the error where the
passed key is unknown, we check that the error stems
from the pkcs11 engine and that the reason is
"object not found". To obtain this information,
SSL_ERR_lib_error_string() and SSL_ERR_reason_error_string()
wrappings are introduced.

Inside HsmEngine::loadPrivateKey() and
HsmEngine::loadPublicKey(), the OpenSSLException
is caught and checked whether it has been raised
due to unknown ID. If so, a MoCOCrWException is
thrown. Otherwise, the original OpenSSLException
is thrown again.