HSM: Detect duplicate key generation
The PKCS#11 standard is unclear on what happens when
a key is already stored in an HSM and an additional generate
operation is invoked to create a new key with the same label
and ID.

Consequently, this change adds a check to MoCOCrW in order to
detect duplicate key generation and raise an exception in such cases.

This is achived by attempting to load the key prior to invoking the
KEYGEN command. If the key is unknown at this point (determined by trying
to catch a MoCOCrWUnknownHsmKeyException), key generation is allowed to
be performed. Otherwise, a MoCOCrWHsmKeyDuplicateGenerationException
is raised.

This change also includes a patch to libp11 which adds a new ENGINE
command to open a RW session, a must for key generation in order to
avoid read-only errors. Prior to the patch, key generation was
could fail if a read-only session was internally used for an earlier
operation.