hsm: don't allow creating keys with the same ID (#140)
* hsm: don't allow creating keys with the same ID

PKCS#11 standard does not enforce unique ids but we want to because
having duplicate ids causes problems on all sides. Before key
generation, key is loaded to check if it already exists. Loading a key
creates a RO session on the token. This represented a problem for key
generation because RW session is needed and libp11 reused previously
created RO session. Key generation patch was slightly modified so that
we always open a RW session before generating a key.