hsm: specify pin before managing keys
Specifying multiple HsmEngine objects in a row overwrites
set pin that is stored in libp11 for the previous HsmEngine objects.
This means that when loading or generating keys, login to the token
fails for all the HsmEngine objects except for the last one defined.
To combat that, pin is stored in HsmEngine object and set in libp11
right before loading/generating keys.