Properly escape PKCS#11 URIs
Our code transforms the token name, key id and key label into a
PKCS#11 URI that is passed into libp11. So far we only escaped the
key id as it naturally consists of non-printable characters.
However, the URIs have some reserved characters that have a special
meaning in the URI. As we currently just forward the token name and
key label to the URI this could lead to accidental or malicious URI
misinterpretation/injection.

In order to mitigate this problem, this patch provides pct encoding
for the special characters. The list of special characters is taken
from RFC 7512 (PKCS#11 URI scheme).