aes: fix iv length resetting for aes-gcm encryption
For AES GCM, iv length is set in EVP_CIPHER_CTX data structure.
This was mistakenly reset to default length of 12 with the calls to
EVP_CIPHER_CTX_reset and EVP_CipherInit_ex. This leads to wrong
ciphertext because only lower 12 bytes of the IV are used.

Example is modified to catch this bug and sanity checks for IV
length are added.