hsm-integration-test: Add python tools
Python tools for integration testing is added.
Basically these are required because the output of pkcs11-tool is not
provided in a machine readable way.

The following tools are added:
* pkcs11_to_json: Reads the output of pkcs11-tool and outputs it in json
  representation
* get_slot_by_label: Takes the output of "pkcs11-tool -L" in a json
  representation and returns the slot ID for a given label.
* key_attribute_validator: Compares the json representation of
  "pkcs11-tool -O ..." of private keys with a given json dict. If the
  content differs, then the script exits with exit code 1

These tools are used for integration tests. The purpose is to check the
key attributes of the private keys created to make sure that
CKA_EXTRACTABLE and CKA_SENSITIVE is set as expected.